+++ internal/mkrpm/mkrpm.go
@@ -156,7 +156,7 @@ func Build(p pkgmeta.Package) (foutpath string, err error) {
 
 	info.Overridables.RPM.Group = p.Group
 
-	if *internal.GPGKeyID != "" {
+	if *internal.GPGKeyPassword != "" {
 		slog.Debug("using GPG key", "file", *internal.GPGKeyFile, "id", *internal.GPGKeyID, "password", *internal.GPGKeyPassword)
 		info.Overridables.RPM.Signature.KeyFile = *internal.GPGKeyFile
 		info.Overridables.RPM.Signature.KeyID = internal.GPGKeyID