+++ CHANGELOG.md
@@ -6,8 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-- Added installation instructions to the `README.md`
 
+- Added installation instructions to the `README.md`
 - Set mtime for deb/rpm package files to unix time 0.
 
 ## v0.0.7